move AppWithMainNav to the src folder
