+++ src/AppWithMainNav.sc.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import {
   sidebarCollapsedWidth,
   sidebarExpandedWidth,
-} from "./MainNav/SideBar/Sidebar.sc";
+} from "./components/sidebar/MainNav/SideBar/Sidebar.sc";
 
 const AppWithMainNav = styled.div`
   box-sizing: border-box;
